+++ src/READUS/Models/CreateAccountRequest.cs
@@ -3,9 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace READUS.Commands
+namespace READUS.Models
 {
-    public class CreateAccountCommand
+    public class CreateAccountRequest
     {
         public string Username { get; set; }
         public string Password { get; set; }
