+++ src/READUS/appsettings.json
@@ -12,7 +12,8 @@
   },
 
   "JWT": {
-    "SymmetricKey": "super_secret_key"
+    "SymmetricKey": "super_secret_key",
+    "TokenLifetimeDays": "7" 
   },
 
     "AllowedHosts": "*"
